Subdivide scene classes
