+++ meerk40t/gui/scene/sceneconst.py
@@ -1,0 +1,24 @@
+
+MILS_IN_MM = 39.3701
+
+HITCHAIN_HIT = 0
+HITCHAIN_DELEGATE = 1
+HITCHAIN_HIT_AND_DELEGATE = 2
+HITCHAIN_DELEGATE_AND_HIT = 3
+
+RESPONSE_CONSUME = 0
+RESPONSE_ABORT = 1
+RESPONSE_CHAIN = 2
+RESPONSE_DROP = 3
+
+ORIENTATION_MODE_MASK = 0b00001111110000
+ORIENTATION_DIM_MASK = 0b00000000001111
+ORIENTATION_MASK = ORIENTATION_MODE_MASK | ORIENTATION_DIM_MASK
+ORIENTATION_RELATIVE = 0b00000000000000
+ORIENTATION_ABSOLUTE = 0b00000000010000
+ORIENTATION_CENTERED = 0b00000000100000
+ORIENTATION_HORIZONTAL = 0b00000001000000
+ORIENTATION_VERTICAL = 0b00000010000000
+ORIENTATION_GRID = 0b00000100000000
+ORIENTATION_NO_BUFFER = 0b00001000000000
+BUFFER = 10.0
