+++ meerk40t/gui/scenewidgets/guidewidget.py
@@ -3,7 +3,7 @@ import math
 import wx
 
 from meerk40t.gui.laserrender import DRAW_MODE_GUIDES
-from meerk40t.gui.scene.scene import Widget
+from meerk40t.gui.scene.widget import Widget
 
 
 class GuideWidget(Widget):
