+++ meerk40t/gui/scenewidgets/laserpathwidget.py
@@ -1,7 +1,7 @@
 import wx
 
 from meerk40t.gui.laserrender import DRAW_MODE_LASERPATH
-from meerk40t.gui.scene.scene import Widget
+from meerk40t.gui.scene.widget import Widget
 
 
 class LaserPathWidget(Widget):
