+++ meerk40t/gui/scenewidgets/reticlewidget.py
@@ -1,7 +1,7 @@
 import wx
 
 from meerk40t.gui.laserrender import DRAW_MODE_RETICLE
-from meerk40t.gui.scene.scene import Widget
+from meerk40t.gui.scene.widget import Widget
 from meerk40t.svgelements import Color
