+++ meerk40t/gui/simulation.py
@@ -2,6 +2,8 @@ import math
 
 import wx
 
+from .scene.scenepanel import ScenePanel
+from .scene.widget import Widget
 from ..core.cutcode import CutCode, LineCut
 from ..kernel import Job, signal_listener
 from ..svgelements import Matrix
@@ -14,7 +16,6 @@ from .icons import (
 from .laserrender import LaserRender
 from .mwindow import MWindow
 from .scenewidgets.gridwidget import GridWidget
-from .scene.scene import ScenePanel, Widget
 from .wxutils import disable_window
 
 _ = wx.GetTranslation
