+++ meerk40t/gui/wxmscene.py
@@ -4,7 +4,7 @@ from wx import aui
 from meerk40t.gui.icons import icon_meerk40t
 from meerk40t.gui.laserrender import LaserRender
 from meerk40t.gui.mwindow import MWindow
-from meerk40t.gui.scene.scene import ScenePanel
+from meerk40t.gui.scene.scenepanel import ScenePanel
 from meerk40t.gui.scenewidgets.elementswidget import ElementsWidget
 from meerk40t.gui.scenewidgets.gridwidget import GridWidget
 from meerk40t.gui.scenewidgets.guidewidget import GuideWidget
